functionality improved

+++ src/main/java/com/example/stocks/web/StockController.java
@@ -4,15 +4,11 @@ import com.example.stocks.dto.StockDTO;
 import com.example.stocks.facade.StockFacade;
 import com.example.stocks.payload.response.MessageResponse;
 import com.example.stocks.service.StockService;
-import com.example.stocks.service.GetStockWikiDataService;
-import com.example.stocks.service.TestingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -27,17 +23,23 @@ public class StockController {
 
 
 	@PostMapping("/{ticker}/create")
-	public ResponseEntity<MessageResponse> createStock(@PathVariable("ticker") String ticker) throws IOException {
+	public ResponseEntity<MessageResponse> createStock(@PathVariable("ticker") String ticker) {
 		stockService.createStock(ticker);
 
 		return new ResponseEntity<>(new MessageResponse("Stock created"), HttpStatus.OK);
 	}
 
+
 	@PostMapping("/{id}/delete")
 	public ResponseEntity<MessageResponse> deleteStock(@PathVariable("id") Long id) {
-		stockService.deleteStock(id);
-		return new ResponseEntity<>(new MessageResponse("Stock deleted"), HttpStatus.OK);
+
+		if (stockService.stockExists(id)) {
+			stockService.deleteStock(id);
+			return new ResponseEntity<>(new MessageResponse("Stock deleted"), HttpStatus.OK);
+		}
+		return new ResponseEntity<>(new MessageResponse("Stock with id [" + id + "] doesn't exist" ), HttpStatus.BAD_REQUEST);
 	}
+
 
 	@GetMapping("/all")
 	public ResponseEntity<List<StockDTO>> getAllStocks() {
@@ -46,6 +48,17 @@ public class StockController {
 				.map(stockFacade::databaseStockToStockDTO)
 				.collect(Collectors.toList());
 		return new ResponseEntity<>(stockDTOList, HttpStatus.OK);
+	}
+
+
+	@PostMapping("/update-stock-info/{id}")
+	public ResponseEntity<MessageResponse> updateStockInfo(@PathVariable("id") Long id,
+														   @RequestParam("companyInfo") String companyInfo) {
+		if (stockService.stockExists(id)) {
+			stockService.updateStockInfo(id, companyInfo);
+			return new ResponseEntity<>(new MessageResponse("Stock info updated"), HttpStatus.OK);
+		}
+		return new ResponseEntity<>(new MessageResponse("No stock found with id [" + id + "]" ), HttpStatus.BAD_REQUEST);
 	}
